refactor(commnds): Refactor commands in seperate files for clarity
