+++ commands/skills/skills.js
@@ -1,0 +1,562 @@
+module.exports = {
+
+  Stun: {
+    title: ['Fatal Blow', 'Power Arrow', 'Magical Shock'],
+    titleFr: ['Coup Fatal', 'Assaut de flèche', 'Choc Magique'],
+    thumbnail: ['Fatal_Blow.png', 'Power_Arrow.png', 'Magical_Shock.png'],
+
+    aliases: ['Stun', 'Fatal Blow', 'Power Arrow', 'Magical Shock', 'Coup Fatal', 'Assaut de flèche', 'Choc Magique', 'H01', 'H07'],
+
+    desc: 'Deal damage and stun the enemy, CRI rate increases when the enemy is stunned.',
+    descFr: 'Inflige des dégâts et étourdit l\'ennemi. Taux de Crit++ lorsque l\'ennemi est assomé.',
+
+    levels: [],
+    grades: ['H01 - 0', 'H07 - 3000'],
+  },
+
+  Berzerk: {
+    title: ['Berserk'],
+    titleFr: ['Fou Furieux'],
+    thumbnail: ['Berserk.png'],
+
+    aliases: ['Berzerk', 'Berserk', 'Fou Furieux', 'H02', 'H10'],
+
+    desc: 'Next Attack will crit and increase control free and crit.',
+    descFr: 'La prochaine ATQ est Crit, contrôle libre et DGT crit augmentés, durée longue',
+
+    levels: [],
+    grades: ['H02 - 200', 'H10 - 6000'],
+  },
+
+  Double: {
+    title: ['Double Attack', 'Double Strafe', 'Magical Double Hit'],
+    titleFr: ['Double Attaque', 'Double Mitrailler', 'Frappe Magic Double'],
+    thumbnail: ['Double_Attack.png', 'Double_Strafe.png', 'Magical_Double_Hit.png'],
+
+    aliases: ['Double', 'Double Attack', 'Double Strafe', 'Magical Double Hit', 'Double Attaque', 'Double Mitrailler', 'Frappe Magic Double', 'H03', 'H14'],
+
+    desc: 'Deals damage twice.',
+    descFr: 'Inflige deux fois des DGT',
+
+    levels: [],
+    grades: ['H03 - 500', 'H14 - 12900'],
+  },
+
+  Frost: {
+    title: ['Frost Blade', 'Frost Bolt', 'Stone Curse'],
+    titleFr: ['Lame de givre', 'Eclair de givre', 'Pierre de Malédiction'],
+    thumbnail: ['Frost_Blade.png', 'Frost_Bolt.png', 'Stone_Curse.png'],
+
+    aliases: ['Frost', 'Frost Blade', 'Frost Bolt', 'Stone Curse', 'Lame de givre', 'Eclair de givre', 'Pierre de Malédiction', 'H04', 'H16'],
+
+    desc: 'Freeze/Stone. Lowers corresponding defence (physical/magic) and raises the opposite defence.',
+    descFr: 'LV2 Warrior : Inflide dégâts et gèle l\'ennemi, réduit Def phys, mais augmente Def Mag',
+
+    levels: [],
+    grades: ['H04 - 1000', 'H16 - 21500'],
+  },
+
+  Provoke: {
+    title: ['Provoke'],
+    titleFr: ['Provoquer'],
+    thumbnail: ['Provoke.png'],
+
+    aliases: ['Provoke', 'Provoke', 'Provoquer', 'H05', 'H19'],
+
+    desc: 'Lowers the enemy\'s DEF and M.DEF significantly.',
+    descFr: 'Réduit la def et def mag. de l\'ennemi de manière significative',
+
+    levels: [],
+    grades: ['H05 - 1500', 'H19 - 32700'],
+  },
+
+  Sleep: {
+    title: ['Sleeping Blade', 'Sleep Bolt', 'Sleep'],
+    titleFr: ['Lame de sommeil', 'Eclair sommeil', 'Sommeil'],
+    thumbnail: ['Sleeping_Blade.png', 'Sleep_Bolt.png', 'Sleep.png'],
+
+    aliases: ['Sleep', 'Sleeping Blade', 'Sleep Bolt', 'Sleep', 'Lame de sommeil', 'Eclair sommeil', 'Sommeil', 'H06', 'H21'],
+
+    desc: 'Fast cast, Deals Dmg, Sleep. More Dmg when the enemy is sleeping.',
+    descFr: 'Inflige des dégâts à l\'ennemi et l\'endort. L\'ennemi recevra plus de dégâts quand il dort.',
+
+    levels: [],
+    grades: ['H06 - 2000', 'H21 - 41800'],
+  },
+
+  'Counter Attack': {
+    title: ['Counter Attack'],
+    titleFr: ['Contre-Attaque'],
+    thumbnail: ['Counter_Attack.png'],
+
+    aliases: ['Counter Attack', 'Counter Attack', 'Contre-Attaque', 'H08', 'H22'],
+
+    desc: 'Bounce the received Dmg, Long period. ',
+    descFr: 'Renvoie les degats subis',
+
+    levels: [],
+    grades: ['H08 - 4000', 'H22 - 46800'],
+  },
+
+  Cure: {
+    title: ['Cure'],
+    titleFr: ['Soin'],
+    thumbnail: ['Cure.png'],
+
+    aliases: ['Cure', 'Cure', 'Soin', 'H09', 'H24'],
+
+    desc: 'Restore your lost HP. And it continues for a period. ',
+    descFr: 'Restaure vos vies perdues',
+
+    levels: [],
+    grades: ['H09 - 5000', 'H24 - 58100'],
+  },
+
+  Silence: {
+    title: ['Blade of Silence', 'Silent Bolt', 'Silence'],
+    titleFr: ['Lame de silence', 'Eclair silencieux', 'Silence'],
+    thumbnail: ['Blade_of_Silence.png', 'Silent_Bolt.png', 'Silence.png'],
+
+    aliases: ['Silence', 'Blade of Silence', 'Silent Bolt', 'Silence', 'Lame de silence', 'Eclair silencieux', 'Silence', 'H11', 'H26'],
+
+    desc: 'Long duration silence but sometimes fails.',
+    descFr: 'Inflige des dégâts a l\'ennemi et le réduit au silence.',
+
+    levels: [],
+    grades: ['H11 - 8500', 'H26 - 70900'],
+  },
+
+  HIT: {
+    title: ['Spear Dynamo', 'Meditation', 'Element Focus'],
+    titleFr: ['Lance dynamo', 'Méditation', 'Elément Cible'],
+    thumbnail: ['Spear_Dynamo.png'],
+
+    aliases: ['HIT', 'Spear Dynamo', 'Meditation', 'Element Focus', 'Lance dynamo', 'Méditation', 'Elément Cible', 'H12', 'H27'],
+
+    desc: 'Improve HIT greatly and it lasts for a long period. ',
+    descFr: 'Améliore grandemt FRAPPE et dure pdt une longue période.',
+
+    levels: [],
+    grades: ['H12 - 10600', 'H27 - 77900'],
+  },
+
+  Dexterity: {
+    title: ['Dexterity', 'Windwalker'],
+    titleFr: ['Dextérité', 'Marcheur de vent'],
+    thumbnail: ['Dexterity.png'],
+
+    aliases: ['Dexterity', 'Dexterity', 'Windwalker', 'Dexterity', 'Dextérité', 'Marcheur de vent', 'Dextérité', 'H14', 'H28'],
+
+    desc: 'Increases EVA. ',
+    descFr: 'Augmente EVIT',
+
+    levels: [],
+    grades: ['H14 - 15500', 'H28 - 85300'],
+  },
+
+  Triple: {
+    title: ['Triple Action', 'Triple Strafe', 'Mental Power'],
+    titleFr: ['Triple Action', 'Triple Tir', 'Pouvoir Mental'],
+    thumbnail: ['Triple_Action.png', 'Triple_Strafe.png', 'Mental_Power.png'],
+
+    aliases: ['Triple', 'Triple Action', 'Triple Strafe', 'Mental Power', 'Triple Action', 'Triple Tir', 'Pouvoir Mental', 'H15', 'H31'],
+
+    desc: 'Improve HIT greatly. And it lasts for a long period. ',
+    descFr: '',
+
+    levels: [],
+    grades: ['H15 - 18300', 'H31 - 101400'],
+  },
+
+  Awaken: {
+    title: ['Awaken'],
+    titleFr: ['Eveiller'],
+    thumbnail: ['Awaken.png'],
+
+    aliases: ['Awaken', 'Awaken', 'Eveiller', 'H17', 'H33'],
+
+    desc: 'Restores your lost SP.',
+    descFr: 'Restore vos SP perdu',
+
+    levels: [],
+    grades: ['H17 - 24900', 'H33 - 121800'],
+  },
+
+  Curse: {
+    title: ['Cursed Blade', 'Curse Bolt', 'Curse Magic'],
+    titleFr: ['Lame de malédiction', 'Eclair de malédiction', 'Malédiction magique'],
+    thumbnail: ['Cursed_Blade.png', 'Curse_Bolt.png', 'Curse_Magic.png'],
+
+    aliases: ['Curse', 'Cursed Blade', 'Curse Bolt', 'Curse Magic', 'Lame de malédiction', 'Eclair de malédiction', 'Malédiction magique', 'H18', 'H35'],
+
+    desc: 'Deals some Dmg, Curse the enemy, Lower enemy\'s HIT and EVA.',
+    descFr: 'Inflige des degats et maudit l\'ennemi. Reduit la FRAPPE et EVIT de l\'ennemi.',
+
+    levels: [],
+    grades: ['H18 - 28600', 'H35 - 139200'],
+  },
+
+  Speed: {
+    title: ['Speed Burst'],
+    titleFr: ['Vitesse Limite'],
+    thumbnail: ['Speed_Burst.png'],
+
+    aliases: ['Speed', 'Speed Burst', 'Vitesse Limite', 'H20', 'H38'],
+
+    desc: 'Greatly increases Atk Spd, Long Duration.',
+    descFr: 'Augmente considérablement la Vitesse d\'Attaque, Durée longue',
+
+    levels: [],
+    grades: ['H20 - 37000', 'H38 - 161000'],
+  },
+
+  H23: {
+    title: ['Furious Blow', 'Fury Arrow', 'Energy Reflection'],
+    titleFr: ['Coup féroce', 'Flèche de Furie', 'Réflexion énergétique'],
+    thumbnail: ['Furious_Blow.png', 'Fury_Arrow.png', 'Energy_Reflection.png'],
+
+    aliases: ['H23', 'Furious Blow', 'Fury Arrow', 'Energy Reflection', 'Coup féroce', 'Flèche de Furie', 'Réflexion énergétique', 'H23', 'H41'],
+
+    desc: 'Deals Dmg, Dmg suffered during cast will increase them, Recovers part of HP lost.',
+    descFr: 'Inflige des dégâts et maudit l\'ennemi. Reduit FRAPPE et EVIT de l\'ennemi',
+
+    levels: [],
+    grades: ['H23 - 52300', 'H41 - 197600'],
+  },
+
+  'AGI Reduction': {
+    title: ['AGI Reduction'],
+    titleFr: ['Diminuer AGI'],
+    thumbnail: ['AGI_Reduction.png'],
+
+    aliases: ['AGI Reduction', 'AGI Reduction', 'Diminuer AGI', 'H25', 'H43'],
+
+    desc: 'Lowers enemy attack speed.',
+    descFr: 'Réduit la vitesse d\'attaque de l\'ennemi pendant une période',
+
+    levels: [],
+    grades: ['H25 - 64300', 'H43 - 224800'],
+  },
+
+  'Target Change': {
+    title: ['Target Change'],
+    titleFr: ['Transfert d\'objet'],
+    thumbnail: ['Target_Change.png'],
+
+    aliases: ['Target Change', 'Target Change', 'Transfert d\'objet', 'H29', 'H37'],
+
+    desc: 'Forces enemies to attack you in teamfights.',
+    descFr: 'Pousse tous les ennemis à vous attaquer mais dure pendant une courte période ',
+
+    levels: [],
+    grades: ['H29 - 90000', 'H37 - 150000'],
+  },
+
+  Concentration: {
+    title: ['Concentration'],
+    titleFr: ['Concentration'],
+    thumbnail: ['Concentration.png'],
+
+    aliases: ['Concentration', 'Concentration', 'Concentration', 'H30', 'H45'],
+
+    desc: 'Reduces duration of debuffs on you.',
+    descFr: 'Durée des altérations d\'état nég. en baisse',
+
+    levels: [],
+    grades: ['H30 - 93100', 'H45 - 254200'],
+  },
+
+  H32: {
+    title: ['Charged Attack', 'Aimed Bolt', 'Element Bomb'],
+    titleFr: ['Assaut', 'Tir ciblé', 'Element Bombe'],
+    thumbnail: ['Charged_Attack.png', 'Aimed_Bolt.png', 'Element_Bomb.png'],
+
+    aliases: ['H32', 'Charged Attack', 'Aimed Bolt', 'Element Bomb', 'Assaut', 'Tir ciblé', 'Element Bombe', 'H32', 'H46'],
+
+    desc: 'High damage attack, also heals you.',
+    descFr: 'Inflige de très grand dégats après la charge, HIT augmenté',
+
+    levels: [],
+    grades: ['H32 - 110200', 'H46 - 269800'],
+  },
+
+  Bleed: {
+    title: ['Vital Strike', 'Falcon Eyes', 'Thunderous Shot'],
+    titleFr: ['Attaque ingénieuse', 'Sniper', 'Tir tonitruant'],
+    thumbnail: ['Vital_Strike.png', 'Falcon_Eyes.png', 'Thunderous_Shot.png'],
+
+    aliases: ['Bleed', 'Vital Strike', 'Falcon Eyes', 'Thunderous Shot', 'Attaque ingénieuse', 'Sniper', 'Tir tonitruant', 'H34', 'H47'],
+
+    desc: 'Deals bonus damage to enemies with a debuff on them.',
+    descFr: 'Inflige des dégats a tous les ennemis. Si ils sont sous altération d\'état, les dégats sont plus grand, et saignement est infligé.',
+
+    levels: [],
+    grades: ['H34 - 129000', 'H47 - 286000'],
+  },
+
+  H36: {
+    title: ['Pierce', 'Ghost Arrow', 'Fire Wall'],
+    titleFr: ['Attaque perçante', 'Flèche Magique', 'Mur de Feu'],
+    thumbnail: ['Pierce.png', 'Ghost_Arrow.png', 'Fire_Wall.png'],
+
+    aliases: ['H36', 'Pierce', 'Ghost Arrow', 'Fire Wall', 'Attaque perçante', 'Flèche Magique', 'Mur de Feu', 'H36', 'H49'],
+
+    desc: 'Deals 3x Incremental Dmg, the 2nd and 3rd hit will not trigger effects.',
+    descFr: 'inflige trois fois des dégats incrémentaux',
+
+    levels: [],
+    grades: ['H36 - 149900', 'H49 - 320300'],
+  },
+
+  H39: {
+    title: ['Traumatic Blow', 'Vulnerability Mark', 'Magical Chase'],
+    titleFr: ['Coup Traumatique', 'Opportuniste', 'Chasse Magique'],
+    thumbnail: ['Traumatic_Blow.png', 'Vulnerability_Mark.png', 'Magical_Chase.png'],
+
+    aliases: ['H39', 'Traumatic Blow', 'Vulnerability Mark', 'Magical Chase', 'Coup Traumatique', 'Opportuniste', 'Chasse Magique', 'H39', 'H50'],
+
+    desc: 'Crit will increase following dmg, Long Duration, will not stack. ',
+    descFr: 'Les crit augmente grandement les dégats ultérieurs, durée longue, non cummulable',
+
+    levels: [],
+    grades: ['H39 - 172700', 'H50 - 338400'],
+  },
+
+  'Life Steal': {
+    title: ['Blade of Blood', 'Blood Steal', 'Energy Transfer'],
+    titleFr: ['Lame de sang', 'Vampire', 'Transfert énergétique'],
+    thumbnail: ['Blade_of_Blood.png', 'Blood_Steal.png', 'Energy_Transfer.png'],
+
+    aliases: ['Life Steal', 'Blade of Blood', 'Blood Steal', 'Energy Transfer', 'Lame de sang', 'Vampire', 'Transfert énergétique', 'H40', 'H51'],
+
+    desc: 'Causes attacks to steal health, also increases crit damage.',
+    descFr: 'Dégats augmenté, et vol de HP, durée moyenne',
+
+    levels: [],
+    grades: ['H40 - 184900', 'H51 - 357100'],
+  },
+
+  Distraction: {
+    title: ['Distraction'],
+    titleFr: ['Distraction'],
+    thumbnail: ['Distraction.png'],
+
+    aliases: ['Distraction', 'Distraction', 'Distraction', 'H42', 'H48'],
+
+    desc: 'Lowers enemy Cast and raises their SP consumption.',
+    descFr: 'Baisse la vitesse d\'invocation et augmente la consommation de SP de l\'ennemi sur une durée longue',
+
+    levels: [],
+    grades: ['H42 - 210900', 'H48 - 302800'],
+  },
+
+  Clearance: {
+    title: ['Clearance'],
+    titleFr: ['Dégagement'],
+    thumbnail: ['Clearance.png'],
+
+    aliases: ['Clearance', 'Clearance', 'Dégagement', 'H44', 'H52'],
+
+    desc: 'Removes enemy buffs, restores your HP if successful, and makes enemies take more damage.',
+    descFr: 'Affaiblis les ennemis, Retire les altérations d\'états pos., Restitue des HP en cas de succès',
+
+    levels: [],
+    grades: ['H44 - 239200', 'H52 - 376500'],
+  },
+
+  'Hidden Potential': {
+    title: ['Hidden Potential'],
+    titleFr: ['Activation de potentiel'],
+    thumbnail: ['Hidden_Potential.png'],
+
+    aliases: ['Hidden Potential', 'Hidden Potential', 'Activation de potentiel', 'H53', 'H61'],
+
+    desc: 'The longer the duration of the fighter, higher is the dmg dealt',
+    descFr: 'Attaque l\'ennemi. Plus le combat est long, plus les DGT augmentent.',
+
+    levels: [],
+    grades: ['H53 - 393300', 'H61 - '],
+  },
+
+  'Power Impact': {
+    title: ['Power Impact'],
+    titleFr: ['Impact continu'],
+    thumbnail: ['Power_Impact.png'],
+
+    aliases: ['Power Impact', 'Power Impact', 'Impact continu', 'H54', 'H57'],
+
+    desc: 'Attacks more times, based on the enemy\'s HP',
+    descFr: 'Attaque l\'ennemi. Plus les HP de l\'ennemi sont éléves, plus les attaques sont nombreuses.',
+
+    levels: [],
+    grades: ['H54 - 400000', 'H57 - '],
+  },
+
+  'Power Shield': {
+    title: ['Power Shield'],
+    titleFr: ['Amortisseur de dégâts'],
+    thumbnail: ['Power_Shield.png'],
+
+    aliases: ['Power Shield', 'Power Shield', 'Amortisseur de dégâts', 'H55', 'H59'],
+
+    desc: '""Delay the dmg taken to affect you""',
+    descFr: 'Les DGT subis durant un certain temps prennet effet plus tard.',
+
+    levels: [],
+    grades: ['H55 - 413300', 'H59 - 478900'],
+  },
+
+  Mislead: {
+    title: ['Mislead'],
+    titleFr: ['Fourvoiement'],
+    thumbnail: ['Mislead.png'],
+
+    aliases: ['Mislead', 'Mislead', 'Fourvoiement', 'H56', 'H63'],
+
+    desc: 'Mislead the Opponent, Causing his skills to affect himself too',
+    descFr: 'Fourvoie l\'ennemi. Sa prochaine compétence aura un effet sur lui-même.',
+
+    levels: [],
+    grades: ['H56 - 432300', 'H63 - '],
+  },
+
+  'Battle Chains': {
+    title: ['Battle Chains'],
+    titleFr: ['Chaînes de combat'],
+    thumbnail: ['Battle_Chains.png'],
+
+    aliases: ['Battle Chains', 'Battle Chains', 'Chaînes de combat', 'H58', 'H64'],
+
+    desc: 'Increases ATK and DEF based on the Combat Pet Attributes',
+    descFr: 'ATQ/DEF ↑ obtenu depuis les attributs de l\'animal de combat.',
+
+    levels: [],
+    grades: ['H58 - ', 'H64 - '],
+  },
+
+  Intimidation: {
+    title: ['Intimidation'],
+    titleFr: ['Intimidation'],
+    thumbnail: ['Intimidation.png'],
+
+    aliases: ['Intimidation', 'Intimidation', 'Intimidation', 'H60', 'H62'],
+
+    desc: 'Critical attacks weaken the enemy, becomes stronger every 5 crits.',
+    descFr: 'Les Crit réduisent les DGT de l\'ennemi, effet renforcé tous les 5 Crit.',
+
+    levels: [],
+    grades: ['H60 - ', 'H62 - '],
+  },
+
+  Bash: {
+    title: ['Bash', 'Brutal Arrow', 'Mindblow'],
+    titleFr: ['Attaque sauvage', 'Flèche Brutale', 'Hallucination'],
+    thumbnail: ['Bash.png', 'Brutal_Arrow.png', 'Mindblow.png'],
+
+    aliases: ['Bash', 'Bash', 'Brutal Arrow', 'Mindblow', 'Attaque sauvage', 'Flèche Brutale', 'Hallucination'],
+
+    desc: 'High damage attack.',
+    descFr: 'Inflige de lourds dégâts',
+
+    levels: [],
+    grades: [''],
+  },
+
+  'Full Throttle': {
+    title: ['Full Throttle'],
+    titleFr: ['Pleins gaz'],
+    thumbnail: ['Full_Throttle.png'],
+
+    aliases: ['Full Throttle', 'Full Throttle', 'Pleins gaz'],
+
+    desc: 'Deals damage and Enhance Stats + class bonus',
+    descFr: 'Inflige des dégâts, Augmente tous les attributs + bonus de classe',
+
+    levels: [],
+    grades: [''],
+  },
+
+  Endure: {
+    title: ['Endure'],
+    titleFr: ['Endure'],
+    thumbnail: ['Endure.png'],
+
+    aliases: ['Endure', 'Endure', 'Endure'],
+
+    desc: 'Reduce received damage and last for a long period.',
+    descFr: 'Réduit dégâts subis et dure pdt une longue période.',
+
+    levels: [],
+    grades: [''],
+  },
+
+  Guardian: {
+    title: ['Guardian'],
+    titleFr: ['Gardien'],
+    thumbnail: ['Guardian.png'],
+
+    aliases: ['Guardian', 'Guardian', 'Gardien'],
+
+    desc: 'Raises accuracy and evasion.',
+    descFr: 'Augmente grandement Frappe et Evit',
+
+    levels: [],
+    grades: [''],
+  },
+
+  'Energy Coat': {
+    title: ['Energy Coat'],
+    titleFr: ['Manteau énergétique'],
+    thumbnail: ['Energy_Coat.png'],
+
+    aliases: ['Energy Coat', 'Energy Coat', 'Manteau énergétique'],
+
+    desc: 'Consume SP to reduce received Dmg. ',
+    descFr: 'Consomme des SP pour réduire les dégats recus.',
+
+    levels: [],
+    grades: [''],
+  },
+
+  'Wind Cutter': {
+    title: ['Wind Cutter'],
+    titleFr: ['Lame venteuse'],
+    thumbnail: ['Wind_Cutter.png'],
+
+    aliases: ['Wind Cutter', 'Wind Cutter', 'Lame venteuse'],
+
+    desc: 'Deals damage and increases enemy SP consumption?',
+    descFr: 'Inflige DGT et peur aux ennemis. Affaiblis les ennemis et augmente leur consommation de SP.',
+
+    levels: [],
+    grades: [''],
+  },
+
+  Toxic: {
+    title: ['Toxic Arrow'],
+    titleFr: ['Flèche Toxique'],
+    thumbnail: ['Toxic_Arrow.png'],
+
+    aliases: ['Toxic', 'Toxic Arrow', 'Flèche Toxique'],
+
+    desc: 'Deals damage, reduces enemy healing, and deals 8% enemy max HP in damage',
+    descFr: 'Inflige des dégats et empoisone les ennemis',
+
+    levels: [],
+    grades: [''],
+  },
+
+  'Jack Frost': {
+    title: ['Jack Frost'],
+    titleFr: ['Jack Frost'],
+    thumbnail: ['Jack_Frost.png'],
+
+    aliases: ['Jack Frost', 'Jack Frost', 'Jack Frost'],
+
+    desc: 'Deals Dmg, Reduces enemy\'s Atk Speed and Cast.',
+    descFr: 'Inflige des dégâts, réduit la vitesse d\'attaque et le Cast de l\'ennemi.',
+
+    levels: [],
+    grades: [''],
+  },
+};
